backend/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.9.3
bleach==6.1.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==2.2.5
flask-cors==6.0.1
fonttools==4.49.0
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.45
googlesearch-python==1.0.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==2.10
ipykernel==6.29.3
ipython==8.22.1
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.11.0
joblib==1.5.1
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.3
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
mysql-connector-python==9.4.0
narwhals==2.0.1
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.1
notebook_shim==0.2.4
numpy==1.24.4
openai==2.3.0
opencv-python==4.12.0.88
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==6.31.1
psutil==5.9.8
pure-eval==0.2.2
pyarrow==21.0.0
pycparser==2.21
pydantic==2.12.0
pydantic_core==2.41.1
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
selenium==4.35.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.47.1
tenacity==9.1.2
terminado==0.18.0
threadpoolctl==3.6.0
tinycss2==1.2.1
toml==0.10.2
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.8.19.20240106
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2024.1
uri-template==1.3.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whois==1.20240129.2
widgetsnbextension==4.0.10
wsproto==1.2.0
